Klikdokter/requirements.txt
# requirements for setup robot
# run this file with
# pip install -r requirements.txt

#dotenv
python-dotenv
slackclient